requirements.txt


cookiecutter
pytz==2018.9
python-slugify==2.0.1
pillow
argon2-cffi==18.3.0
redis<3,>=2.10.6
django==2.0.10
django-environ==0.4.5
django-model-utils==3.1.2
django-allauth==0.38.0
django-crispy-forms==1.7.2
django-redis==4.10.0
djangorestframework==3.9.0
coreapi==2.3.3
werkzeug==0.14.1
ipdb==0.11
sphinx==1.8.3
psycopg2-binary==2.7.6.1
mypy==0.650
pytest==4.1.1
pytest-sugar==0.9.2
flake8==3.6.0
coverage==4.5.2
factory-boy==2.11.1
django-debug-toolbar==1.11
django-extensions==2.1.4
django-coverage-plugin==1.6.0
pytest-django==3.4.5
django-imagekit
python-dotenv
graphene-django
django-graphql-jwt
autopep8
pylint
django-taggit
django-taggit-serializer
django-cors-headers
boto3
twilio
channels
asgi-redis
geopy
environ

autopep8
yapf
black
